requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
pydantic[email]
pydantic[dotenv]
pydantic-settings
alembic
passlib[bcrypt]
python-jose[cryptography]
asyncpg
psycopg2-binary
greenlet